Add "system-component-addon" category icon
To attempt to fix a missing "Indicators" icon in MATE Control Centre.

This one was a bit weird, since the .desktop file in
"/usr/share/applications" hardcodes the icon to
"/usr/share/icons/hicolor/scalable/apps/ayatana-settings.svg", but for
whatever reason, the one in MATE Control Centre expects an icon named
"system-component-addon".

In my testing, this didn't work, but no harm in adding a new category
icon.